desafio do 1° ao 7° passando e pastas com nome certo agora AMEM

+++ desafio1.sql
@@ -7,7 +7,7 @@ USE SpotifyClone;
 CREATE TABLE planos(
 id INT PRIMARY KEY auto_increment,
 nome_plano VARCHAR(50) NOT NULL,
-valor DOUBLE NOT NULL
+valor DECIMAL(5, 2) NOT NULL
 )ENGINE=InnoDB;
 
 CREATE TABLE usuario(
